+++ packages/docs/react-docs/src/routes/components+/alert-banner+/demos/alert-banner-variant-demo.tsx
@@ -1,0 +1,17 @@
+import type {ReactElement} from "react"
+
+import {AlertBanner} from "@qualcomm-ui/react/alert-banner"
+
+export function AlertBannerVariantDemo(): ReactElement {
+  return (
+    <div className="grid w-full gap-4">
+      {/* preview */}
+      <AlertBanner emphasis="info" heading="info" variant="subtle" />
+      <AlertBanner emphasis="success" heading="success" variant="subtle" />
+      <AlertBanner emphasis="warning" heading="warning" variant="subtle" />
+      <AlertBanner emphasis="danger" heading="danger" variant="subtle" />
+      <AlertBanner emphasis="neutral" heading="neutral" variant="subtle" />
+      {/* preview */}
+    </div>
+  )
+}