+++ packages/docs/react-docs/src/routes/components+/alert-banner+/demos/index.ts
@@ -1,0 +1,6 @@
+export * from "./alert-banner-action-demo"
+export * from "./alert-banner-composite-demo"
+export * from "./alert-banner-dismissable-demo"
+export * from "./alert-banner-emphasis-demo"
+export * from "./alert-banner-simple-demo"
+export * from "./alert-banner-variant-demo"