+++ packages/frameworks/angular/src/alert-banner/alert-banner.module.ts
@@ -1,0 +1,38 @@
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import {NgModule} from "@angular/core"
+
+import {IconDirective} from "@qualcomm-ui/angular/icon"
+import {QBindDirective} from "@qualcomm-ui/angular-core/machine"
+
+import {AlertBannerActionDirective} from "./alert-banner-action.directive"
+import {AlertBannerCloseButtonDirective} from "./alert-banner-close-button.directive"
+import {AlertBannerDescriptionDirective} from "./alert-banner-description.directive"
+import {AlertBannerHeadingDirective} from "./alert-banner-heading.directive"
+import {AlertBannerIconDirective} from "./alert-banner-icon.directive"
+import {AlertBannerRootDirective} from "./alert-banner-root.directive"
+import {AlertBannerDirective} from "./alert-banner.directive"
+
+@NgModule({
+  declarations: [
+    AlertBannerDirective,
+    AlertBannerRootDirective,
+    AlertBannerActionDirective,
+    AlertBannerCloseButtonDirective,
+    AlertBannerDescriptionDirective,
+    AlertBannerHeadingDirective,
+    AlertBannerIconDirective,
+  ],
+  exports: [
+    AlertBannerDirective,
+    AlertBannerRootDirective,
+    AlertBannerActionDirective,
+    AlertBannerCloseButtonDirective,
+    AlertBannerDescriptionDirective,
+    AlertBannerHeadingDirective,
+    AlertBannerIconDirective,
+  ],
+  imports: [IconDirective, QBindDirective],
+})
+export class AlertBannerModule {}