+++ packages/frameworks/angular/src/avatar/avatar.directive.ts
@@ -12,6 +12,7 @@ import type {SignalifyInput} from "@qualcomm-ui/angular-core/signals"
 import {
   createQdsAvatarApi,
   type QdsAvatarApiProps,
+  type QdsAvatarEmphasis,
   type QdsAvatarSize,
   type QdsAvatarStatus,
   type QdsAvatarVariant,
@@ -32,6 +33,11 @@ export class AvatarDirective
   implements SignalifyInput<QdsAvatarApiProps>
 {
   /**
+   * Governs the color of the avatar.
+   */
+  readonly emphasis = input<QdsAvatarEmphasis | undefined>()
+
+  /**
    * Governs the width and height of the avatar as well as the font size of its
    * content.
    */
@@ -43,7 +49,7 @@ export class AvatarDirective
   readonly status = input<QdsAvatarStatus | undefined>()
 
   /**
-   * The style variant of the avatar.
+   * @deprecated use {@link emphasis} instead
    */
   readonly variant = input<QdsAvatarVariant | undefined>()
 
@@ -62,6 +68,7 @@ export class AvatarDirective
     const qdsAvatarApi = computed(() =>
       createQdsAvatarApi(
         {
+          emphasis: this.emphasis(),
           size: this.size(),
           status: this.status(),
           variant: this.variant(),