+++ packages/frameworks/react/src/alert-banner/alert-banner-action-container.tsx
@@ -1,0 +1,39 @@
+// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+
+import type {ReactElement, ReactNode} from "react"
+
+import {
+  type ElementRenderProp,
+  PolymorphicElement,
+} from "@qualcomm-ui/react-core/system"
+import {mergeProps} from "@qualcomm-ui/utils/merge-props"
+
+import {useAlertBannerContext} from "./qds-alert-banner-context"
+
+export interface AlertBannerActionContainerProps
+  extends ElementRenderProp<"div"> {
+  /**
+   * React {@link https://react.dev/learn/passing-props-to-a-component#passing-jsx-as-children children} prop.
+   */
+  children?: ReactNode
+}
+
+/**
+ * A container for the banner's primary action button. Renders a `<div>`
+ * element by default.
+ */
+export function AlertBannerActionContainer({
+  children,
+  ...props
+}: AlertBannerActionContainerProps): ReactElement {
+  const context = useAlertBannerContext()
+
+  const mergedProps = mergeProps(context.getActionBindings(), props)
+
+  return (
+    <PolymorphicElement as="div" {...mergedProps}>
+      {children}
+    </PolymorphicElement>
+  )
+}