+++ packages/frameworks/angular/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @qualcomm-ui/angular Changelog
 
+## 1.17.0 (2026/02/11)
+
+### Features
+
+- alert banner
+- [select]: add checkbox option
+
+### Bug Fixes
+
+- [avatar]: `variant` should be `emphasis`
+- inline notification
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/qds-core]
+
 ## 1.16.0 (2026/01/30)
 
 ### Features